Jan 17, 2023: Redesign all pages
A colors objects was created to organize the application colors. These
colors were added to the WelcomeScreen.js and the rest of the pages.

BackgroundView was added so that I am able to add a custom background to
the homepage.

Added a AppDetails.json in order to store application data so that I
don't have to worry about changing the title in every place the title
is used or etc... For right now we have only the title store in this
json object.

Shapes.js was added but may be remove if I don't find any use for it.